feat(vscode): expose custom agent permissions
