+++ .changeset/per-agent-permissions.md
@@ -1,0 +1,5 @@
+---
+"kilo-code": patch
+---
+
+Support configuring custom agent tool permissions from the VS Code agent editor.
