+++ packages/kilo-vscode/webview-ui/src/stories/settings.stories.tsx
@@ -102,6 +102,14 @@ export const AgentBehaviourEditCustomMode: Story = {
         prompt: "You are a code reviewer. Focus on code quality, best practices, and potential bugs.",
         model: "anthropic/claude-sonnet-4-20250514",
         temperature: 0.3,
+        permission: {
+          read: "allow",
+          grep: "allow",
+          glob: "allow",
+          edit: "deny",
+          bash: "deny",
+          task: "ask",
+        },
       },
     }
     return (
@@ -314,6 +322,51 @@ export const ModeEditExport: Story = {
       <StoryProviders sessionID="export-story" status="idle" config={{ agent: cfg } as any}>
         <SessionContext.Provider value={session as any}>
           <div style={{ width: "420px", height: "700px", overflow: "auto" }}>
+            <ModeEditView name="reviewer" onBack={noop} onRemove={noop} />
+          </div>
+        </SessionContext.Provider>
+      </StoryProviders>
+    )
+  },
+}
+
+export const ModeEditPermissions: Story = {
+  name: "ModeEditView — per-agent permissions",
+  render: () => {
+    const cfg: Record<string, AgentConfig> = {
+      reviewer: {
+        description: "Review code without editing it",
+        prompt: "Find bugs, regressions, and missing tests.",
+        permission: {
+          "*": "deny",
+          read: "allow",
+          grep: "allow",
+          glob: "allow",
+          edit: { "*": "deny", "**/*.md": "allow" },
+          bash: "deny",
+          task: "ask",
+          skill: "deny",
+        },
+      },
+    }
+    const session = {
+      ...mockSessionValue({ id: "permissions-story", status: "idle" }),
+      agents: () => MOCK_AGENTS,
+      allAgents: () => MOCK_AGENTS,
+      removeMode: noop,
+      removeMcp: noop,
+      skills: () => [],
+      refreshSkills: noop,
+      removeSkill: noop,
+    }
+    return (
+      <StoryProviders
+        sessionID="permissions-story"
+        status="idle"
+        config={{ permission: { bash: "ask", external_directory: "ask" }, agent: cfg } as any}
+      >
+        <SessionContext.Provider value={session as any}>
+          <div style={{ width: "460px", height: "760px", overflow: "auto" }}>
             <ModeEditView name="reviewer" onBack={noop} onRemove={noop} />
           </div>
         </SessionContext.Provider>
