+++ packages/kilo-vscode/webview-ui/src/types/messages/permissions.ts
@@ -1,7 +1,7 @@
 export type PermissionLevel = "allow" | "ask" | "deny"
 
 /** null in a PermissionRule object is a delete sentinel — removes the key from the config */
-export type PermissionRule = PermissionLevel | Record<string, PermissionLevel | null>
+export type PermissionRule = PermissionLevel | null | Record<string, PermissionLevel | null>
 
 export type PermissionConfig = Partial<Record<string, PermissionRule>>
